Admin routes for blog publishing and overall app settings
