+++ src/controllers/api/commentController.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from "express";
-import { PrismaClient } from "../../generated/prisma";
-import idSchema from "../schemas/idSchema";
+import { PrismaClient } from "../../../generated/prisma";
+import idSchema from "../../schemas/idSchema";
 import {
   createCommentSchema,
   editCommentSchema,
-} from "../schemas/commentSchema";
+} from "../../schemas/commentSchema";
 
 const prisma = new PrismaClient();
