+++ src/controllers/api/userController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { PrismaClient } from "../../generated/prisma";
+import { PrismaClient } from "../../../generated/prisma";
 
 const prisma = new PrismaClient();
