+++ src/routes/admin/settingsRoutes.ts
@@ -1,0 +1,5 @@
+import { Router } from "express";
+
+const router = Router();
+
+export { router as settingsRouter };
